# Expose Regex Escape & Unescape Span overloads