--- a/src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/CharCategory.cs
+++ b/src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/CharCategory.cs
@@ -4,7 +4,7 @@
 
 namespace System.Text.RegularExpressions
 {
-    internal class RegexCharCategory
+    internal class CharCategory
     {
         private const byte Q = 5;    // quantifier
         private const byte S = 4;    // ordinary stopper
@@ -15,7 +15,8 @@ namespace System.Text.RegularExpressions
         /// <summary>
         /// For categorizing ASCII characters.
         /// </summary>
-        private static readonly byte[] s_category = new byte[] {
+        private static ReadOnlySpan<byte> Category => new byte[]
+        {
             // 0 1 2 3 4 5 6 7 8 9 A B C D E F 0 1 2 3 4 5 6 7 8 9 A B C D E F
                0,0,0,0,0,0,0,0,0,X,X,0,X,X,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
             //   ! " # $ % & ' ( ) * + , - . / 0 1 2 3 4 5 6 7 8 9 : ; < = > ?
@@ -23,31 +24,32 @@ namespace System.Text.RegularExpressions
             // @ A B C D E F G H I J K L M N O P Q R S T U V W X Y Z [ \ ] ^ _
                0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,S,S,0,S,0,
             // ' a b c d e f g h i j k l m n o p q r s t u v w x y z { | } ~
-               0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,Q,S,0,0,0};
+               0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,Q,S,0,0,0
+        };
 
         /// <summary>
         /// Returns true for those characters that terminate a string of ordinary chars.
         /// </summary>
-        public static bool IsSpecial(char ch) => (ch <= '|' && s_category[ch] >= S);
+        public static bool IsSpecial(char ch) => (ch <= '|' && Category[ch] >= S);
 
         /// <summary>
         /// Returns true for those characters that terminate a string of ordinary chars.
         /// </summary>
-        public static bool IsStopperX(char ch) => (ch <= '|' && s_category[ch] >= X);
+        public static bool IsStopperX(char ch) => (ch <= '|' && Category[ch] >= X);
 
         /// <summary>
         /// Returns true for those characters that begin a quantifier.
         /// </summary>
-        public static bool IsQuantifier(char ch) => (ch <= '{' && s_category[ch] >= Q);
+        public static bool IsQuantifier(char ch) => (ch <= '{' && Category[ch] >= Q);
 
         /// <summary>
         /// Returns true for whitespace.
         /// </summary>
-        public static bool IsSpace(char ch) => (ch <= ' ' && s_category[ch] == X);
+        public static bool IsSpace(char ch) => (ch <= ' ' && Category[ch] == X);
 
         /// <summary>
         /// Returns true for chars that should be escaped.
         /// </summary>
-        public static bool IsMetachar(char ch) => (ch <= '|' && s_category[ch] >= E);
+        public static bool IsMetachar(char ch) => (ch <= '|' && Category[ch] >= E);
     }
 }