… src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCharCategory.cs
… src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexOptions.cs
… src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexParser.cs
… src/System.Text.RegularExpressions/tests/Regex.EscapeUnescape.Tests.cs
… src/System.Text.RegularExpressions/tests/SpanTestHelper.cs