Unfortunately using the common path for the Escape string version decreases performance and allocation significantly. I'll probably duplicate code paths to avoid the performance drop. Please notice the performance and huge allocation improvements for Unescape.

Adding a minimal parser for escape / unescape. The end goal is to converge the RegexParser and the ParserMin back into one but that requires collection support for ReadOnlySpan<char> lookups.

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763
Intel Core i7-6600U CPU 2.60GHz (Max: 2.61GHz) (Skylake), 1 CPU, 4 logical and 2 physical cores
  [Host] : .NET Core ? (CoreCLR 4.6.27021.03, CoreFX 4.6.27121.0), 64bit RyuJIT

Job=.NET Core 3.0 regex  Runtime=Core  Toolchain=InProcessToolchain  

```

## Before

|        Method |      Mean |    Error |    StdDev |      Gen 0 | Allocated |
|-------------- |----------:|---------:|----------:|-----------:|----------:|
|   RegexEscape(string) |  93.35 ms | 1.745 ms | 1.7920 ms | 35333.3333 |   70.8 MB |
| RegexUnescape(string) | 106.74 ms | 1.020 ms | 0.9538 ms | 51200.0000 | 102.54 MB |

## After

|        Method |      Mean |    Error |   StdDev |    Median |      Gen 0 | Allocated |
|-------------- |----------:|---------:|---------:|----------:|-----------:|----------:|
|   RegexEscape(string) | 118.22 ms | 2.875 ms | 2.689 ms | 118.09 ms | 37800.0000 |  75.68 MB |
| RegexUnescape(string) |  99.92 ms | 1.931 ms | 1.983 ms |  98.48 ms | 18800.0000 |  37.84 MB |