@danmosemsft 2018-10-23
Nice improvements. This relates to https://github.com/dotnet/corefx/issues/24145 which wasn't formally "approved". Or maybe this part of the API was?