@karelz 2018-12-19
@ViktorHofer what are plans for this PR? Looks like not much progress in last 2 weeks ...