@stephentoub 2019-01-02
I'm closing this for now as:
- There hasn't been any activity here in over a month.
- This is adding new APIs that, as far as I can tell, haven't been approved.